requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
filelock==3.9.0
flower==1.2.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
oauthlib==3.2.2
optional-django==0.3.0
packaging==23.0
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
pycparser==2.21
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
pytest-watch==4.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
react==4.3.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
watchdog==2.2.1
wcwidth==0.2.6
